Simplify concat example
The concat example in the docs assumes that data_a and data_b are
both word-aligned 32-byte. The message is always 64-bytes.

So the circuit should just check that each wire of the message is equal
to the corresponding wire in data_a or data_b. The len_bytes wire is not
requried.