remove "Data Structures" for now (#84)
the only example it gave was an example of concatenating two bundles of wires, of fixed length, by constraining that a double-length array be equal to the concatenation of the first two wire-wise. but this completely misses the point, since really we should just concatenate the original two wire arrays and be done with it---no constraints. the only nontrivial case of concatenation is of variable-length inputs, but this was just only mentioned in one sentence. somebody should come back and do this with variable-length inputs.